rename swarming mod to p2p, fix warnings, update architecture
